backend-python/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
cassandra-driver==3.29.0
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Messaging & Events
aiokafka==0.10.0
aio-pika==9.3.1

# IoT Protocols
paho-mqtt==1.6.1
aiocoap==0.4.7
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
orjson==3.9.10
httpx==0.26.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# WebSockets
python-socketio==5.11.0
websockets==12.0

# Background tasks
celery==5.3.4
flower==2.0.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
isort==5.13.2
mypy==1.8.0
